Build musl-gcc for all architectures we support
The following cross compiles the musl-gcc for all the architectures that
we want to support from that one architecture. For example, we'll build
amd64, arm64, ppc64le and s390x for both amd64 and arm64. Both amd64 and
arm64 are numerous in our CI pipeline, but the other two are not.
They're only around to enable cross compilation.